+++ src/parallelization/hiparsimulator/test/parallel_mpi_9/partitionmanagertest.h
@@ -28,7 +28,7 @@ public:
         // the StripingPartition.
         offset = 130 * dimensions.x();
         ghostZoneWidth = 6;
-        weights = SuperVector<unsigned>(layer.size(), 30 * dimensions.x());
+        weights = SuperVector<long>(layer.size(), 30 * dimensions.x());
         weights[3] = 40 * dimensions.x();
         weights[5] = 20 * dimensions.x();
         // sanity check
@@ -168,7 +168,7 @@ private:
     PartitionManager<2> manager;
     StripingPartition<2> partition;
     Coord<2> dimensions;
-    SuperVector<unsigned> weights;
+    SuperVector<long> weights;
     unsigned offset;
     unsigned ghostZoneWidth;
